add .net framework 4 and 32bit support. it uses on centos5 and winxp with mono.

+++ dnYara.Interop_ancient/Interops/Types/YR_META.cs
@@ -15,9 +15,9 @@ namespace dnYara.Interop
         /// <summary>
         /// Meta identifier.
         /// </summary>
-        public IntPtr identifier;
+        public long identifier;
 
-        public IntPtr strings;
+        public long strings;
 
         public long integer;
